requirements.txt
flask
openai
python-dotenv


# Make chatbot
# Hold previous memory